feat(17-04): migrate Student + User lists to shared DataList
Plan D wave 3, surfaces 4-5 of 5 — the highest-gotcha pair.

- StudentList: parent <div data-testid="student-table"> wraps the DataList
  (Pattern S3 wrapper preserved). getRowTestId='student-row-${id}' applies
  to BOTH desktop tr AND mobile-card wrapper (DataList.tsx:105+148),
  unifying the legacy desktop `student-row-${id}` and mobile
  `student-card-${id}` selectors. data-row-action stopPropagation on
  checkbox + DropdownMenu cells (Pattern S4). Mobile checkbox stays in an
  h-11 w-11 <label> for the 44px hit-target floor (Pattern S1). Header-row
  select-all checkbox lifted out of <th> into a small toolbar above the
  list (DataList does not own header-action columns).
- UserList: pagination block (UserListTable.tsx:188-231) split out of the
  list component and relocated as adjunct in users.index.tsx — DataList
  does not own pagination per CONTEXT D-09. Loading + empty branches fold
  into DataList's `loading` (5-row skeleton) + `emptyState` props.
- 4 OLD files deleted: StudentListTable, StudentMobileCards,
  UserListTable, UserMobileCards.
- 2 routes refactored: students.index.tsx single-import StudentList;
  users.index.tsx single-import UserList + adjunct pagination.
- E2E spec admin-students-crud.mobile.spec.ts updated to query the unified
  `student-row-` test-id prefix (DataList applies same testid to both
  branches, so the legacy `student-card-` prefix no longer renders).
- All 5 admin list surfaces now use sm: (640px) breakpoint convention
  via DataList; the four md: surfaces (Teacher/Student/Subject/User)
  have joined Class on sm:.
